requirements/base.txt
begins >= 0.9
ipython >= 6.2.1
pyglet >= 1.3.1
sunvox-dll-python >= 0.2.0.1.9.3.1
